+++ src/_view/awaitable.c
@@ -180,7 +180,7 @@ gen_next(PyObject *self)
                         Py_None);
         return NULL;
     }
-    
+
     if (g->gw_current_await == NULL) {
         cb = aw->aw_callbacks[aw->aw_state++];
 
@@ -312,7 +312,7 @@ awaitable_dealloc(PyObject *self)
     for (int i = 0; i < aw->aw_callback_size; i++) {
         awaitable_callback *cb = aw->aw_callbacks[i];
         if (!cb->done) Py_DECREF(cb->coro);
-        free(cb);
+        PyMem_Free(cb);
     }
 
     if (aw->aw_arb_values) PyMem_Free(aw->aw_arb_values);
@@ -566,7 +566,7 @@ PyAwaitable_AddAwait(
     Py_INCREF(aw);
     PyAwaitableObject *a = (PyAwaitableObject *) aw;
 
-    awaitable_callback *aw_c = malloc(sizeof(awaitable_callback));
+    awaitable_callback *aw_c = PyMem_Malloc(sizeof(awaitable_callback));
     if (aw_c == NULL) {
         Py_DECREF(aw);
         Py_DECREF(coro);
@@ -588,7 +588,7 @@ PyAwaitable_AddAwait(
         --a->aw_callback_size;
         Py_DECREF(aw);
         Py_DECREF(coro);
-        free(aw_c);
+        PyMem_Free(aw_c);
         PyErr_NoMemory();
         return -1;
     }